--- projects/ADbQuery/test/analyze_function_expression.test.js
@@ -1,6 +1,0 @@
-import * as parser from "../src/parser.js";
-import * as analyzer from "../src/analyzer.js";
-
-test("assignment to field", () => {});
-
-test("assignment from argument", () => {});
